requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
numpy==1.19.0
tensorflow==2.2.0
Pillow==7.2.0
streamlit==0.71.0